# docs/swagger.yaml
definitions:
...
  model.Info:
    properties:
      ID:
        description: 主键ID
        type: integer
      attachments:
        description: 附件
        items:
          type: object
        type: array
      content:
        description: 内容
        type: string
      createdAt:
        description: 创建时间
        type: string
      title:
        description: 标题
        type: string
      updatedAt:
        description: 更新时间
        type: string
      userID:
        description: 作者
        type: integer
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.7.0
paths:
...
  /info/createInfo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建公告
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Info'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建公告
      tags:
      - Info
  /info/deleteInfo:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除公告
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Info'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除公告
      tags:
      - Info
  /info/deleteInfoByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除公告
      tags:
      - Info
  /info/findInfo:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 内容
        in: query
        name: content
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 标题
        in: query
        name: title
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - description: 作者
        in: query
        name: userID
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    reinfo:
                      $ref: '#/definitions/model.Info'
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询公告
      tags:
      - Info
...
  /info/getInfoList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取公告列表
      tags:
      - Info
  /info/getInfoPublic:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 不需要鉴权的公告接口
      tags:
      - Info
  /info/updateInfo:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新公告
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Info'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新公告
      tags:
      - Info
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"